fix fov in camera

+++ raytracer/src/lib/camera/PerspectiveCamera.h
@@ -21,5 +21,5 @@ private:
 
 	Point origin;
 	OrthonormalBasis basis;
-	int width;
+	double width;
 };
